refactor(menu): split components and update style
