+++ src/components/Menu/index.tsx
@@ -1,36 +1,15 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
-import { Trans, t } from '@lingui/macro'
-import React, { useRef } from 'react'
+import { Trans } from '@lingui/macro'
+import { useRef } from 'react'
 import { isMobile } from 'react-device-detect'
-import {
-  Award,
-  BookOpen,
-  Edit,
-  FileText,
-  Info,
-  Menu as MenuIcon,
-  MessageCircle,
-  PieChart,
-  Share2,
-  Triangle,
-  UserPlus,
-} from 'react-feather'
-import { NavLink } from 'react-router-dom'
-import { useMedia } from 'react-use'
-import { Text } from 'rebass'
+import { Menu as MenuIcon } from 'react-feather'
+import { Flex, Text } from 'rebass'
 import styled, { css } from 'styled-components'
 
-import { ReactComponent as BlogIcon } from 'assets/svg/blog.svg'
-import { ReactComponent as LightIcon } from 'assets/svg/light.svg'
-import { ReactComponent as RoadMapIcon } from 'assets/svg/roadmap.svg'
 import { ButtonPrimary } from 'components/Button'
-import { SlideToUnlock } from 'components/Header'
-import DiscoverIcon from 'components/Icons/DiscoverIcon'
-import Faucet from 'components/Icons/Faucet'
 import Loader from 'components/Loader'
 import MenuFlyout from 'components/MenuFlyout'
 import Preferences from 'components/Preferences'
-import { AGGREGATOR_ANALYTICS_URL, DMM_ANALYTICS_URL } from 'constants/index'
 import { NETWORKS_INFO } from 'constants/networks'
 import { useActiveWeb3React } from 'hooks'
 import useClaimReward from 'hooks/useClaimReward'
@@ -38,54 +17,20 @@ import useMixpanel, { MIXPANEL_TYPE } from 'hooks/useMixpanel'
 import useTheme from 'hooks/useTheme'
 import { ApplicationModal } from 'state/application/actions'
 import { useModalOpen, useToggleModal } from 'state/application/hooks'
-import { ExternalLink } from 'theme'
 
 import ClaimRewardModal from './ClaimRewardModal'
 import FaucetModal from './FaucetModal'
-import NavDropDown from './NavDropDown'
+import MenuItems from './MenuItems'
 
-const sharedStylesMenuItem = css`
-  flex: 1;
-  padding: 0.75rem 0;
-  text-decoration: none;
-  display: flex;
-  font-weight: 500;
-  white-space: nowrap;
-  align-items: center;
-  color: ${({ theme }) => theme.subText};
-
-  :hover {
-    color: ${({ theme }) => theme.text};
-    cursor: pointer;
-    text-decoration: none;
-  }
-
-  > svg {
-    margin-right: 8px;
-  }
+const Divider = styled.div`
+  width: 100%;
+  height: 0;
+  border-bottom: 1px solid ${({ theme }) => theme.border};
 `
 
 const StyledMenuIcon = styled(MenuIcon)`
   path {
     stroke: ${({ theme }) => theme.text};
-  }
-`
-
-const StyledRoadMapIcon = styled(RoadMapIcon)`
-  path {
-    stroke: ${({ theme }) => theme.subText};
-  }
-`
-
-const StyledBlogIcon = styled(BlogIcon)`
-  path {
-    fill: ${({ theme }) => theme.subText};
-  }
-`
-
-const StyledLightIcon = styled(LightIcon)`
-  path {
-    stroke: ${({ theme }) => theme.subText};
   }
 `
 
@@ -129,34 +74,11 @@ const StyledMenu = styled.div`
   text-align: left;
 `
 
-export const NavMenuItem = styled(NavLink)`
-  ${sharedStylesMenuItem}
-`
-
-export const ExternalNavMenuItem = styled(ExternalLink)`
-  ${sharedStylesMenuItem}
-`
-
-const MenuButton = styled.div`
-  ${sharedStylesMenuItem}
-`
-
 const MenuFlyoutBrowserStyle = css`
   min-width: 17rem;
-  right: -8px;
-
-  & ${ExternalNavMenuItem}:nth-child(1),
-  & ${NavMenuItem}:nth-child(1) {
-    padding-top: 0.75rem;
-  }
+  right: -11px;
 `
 
-const MenuFlyoutMobileStyle = css`
-  & ${ExternalNavMenuItem}:nth-child(1),
-  & ${NavMenuItem}:nth-child(1) {
-    padding-top: 0.75rem;
-  }
-`
 const ClaimRewardButton = styled(ButtonPrimary)`
   margin-top: 20px;
   padding: 11px;
@@ -171,26 +93,14 @@ export const NewLabel = styled.span`
   margin-left: 2px;
 `
 
-export default function Menu() {
+const Menu: React.FC = () => {
   const { chainId, account } = useActiveWeb3React()
   const theme = useTheme()
   const node = useRef<HTMLDivElement>()
   const open = useModalOpen(ApplicationModal.MENU)
   const toggle = useToggleModal(ApplicationModal.MENU)
 
-  const under1440 = useMedia('(max-width: 1440px)')
-  const above1321 = useMedia('(min-width: 1321px)')
-  const above768 = useMedia('(min-width: 768px)')
-  const under369 = useMedia('(max-width: 370px)')
-
-  const getBridgeLink = () => {
-    if (!chainId) return ''
-    return NETWORKS_INFO[chainId].bridgeURL
-  }
-
-  const bridgeLink = getBridgeLink()
   const toggleClaimPopup = useToggleModal(ApplicationModal.CLAIM_POPUP)
-  const toggleFaucetPopup = useToggleModal(ApplicationModal.FAUCET_POPUP)
   const { pendingTx } = useClaimReward()
   const { mixpanelHandler } = useMixpanel()
   return (
@@ -199,167 +109,53 @@ export default function Menu() {
         <StyledMenuIcon />
       </StyledMenuButton>
 
-      <MenuFlyout
-        node={node}
-        browserCustomStyle={MenuFlyoutBrowserStyle}
-        mobileCustomStyle={MenuFlyoutMobileStyle}
-        isOpen={open}
-        toggle={toggle}
-        hasArrow
-      >
-        <Text fontWeight={500} fontSize={16} color={theme.text}>
-          {t`Menu`}
-        </Text>
-
-        {chainId && [ChainId.BTTC, ChainId.RINKEBY].includes(chainId) && (
-          <MenuButton
-            onClick={() => {
-              toggleFaucetPopup()
-              mixpanelHandler(MIXPANEL_TYPE.FAUCET_MENU_CLICKED)
-            }}
-          >
-            <Faucet />
-            <Text width="max-content">
-              <Trans>Faucet</Trans>
-            </Text>
-          </MenuButton>
-        )}
-
-        {bridgeLink && (
-          <ExternalNavMenuItem href={bridgeLink}>
-            <Share2 size={14} />
-            <Text width="max-content">
-              <Trans>Bridge Assets</Trans>
-            </Text>
-          </ExternalNavMenuItem>
-        )}
-
-        {!above768 && (
-          <NavMenuItem to={'/discover?tab=trending_soon'} onClick={toggle}>
-            <DiscoverIcon size={14} />
-            <SlideToUnlock>
-              <Text width="max-content">
-                <Trans>Discover</Trans>
-              </Text>
-            </SlideToUnlock>
-            <NewLabel>
-              <Trans>New</Trans>
-            </NewLabel>
-          </NavMenuItem>
-        )}
-
-        {under369 && (
-          <NavMenuItem to="/campaigns" onClick={toggle}>
-            <Award size={14} />
-            <Trans>Campaigns</Trans>
-          </NavMenuItem>
-        )}
-        {under1440 && (
-          <NavDropDown
-            icon={<Info size={14} />}
-            title={'About'}
-            link={'/about'}
-            options={[
-              { link: '/about/kyberswap', label: 'Kyberswap' },
-              { link: '/about/knc', label: 'KNC' },
-            ]}
-          />
-        )}
-
-        <NavMenuItem to="/referral" onClick={toggle}>
-          <UserPlus size={14} />
-          <Trans>Referral</Trans>
-        </NavMenuItem>
-
-        {!above1321 && (
-          <NavDropDown
-            icon={<PieChart size={14} />}
-            link="#"
-            title={'Analytics'}
-            options={[
-              { link: DMM_ANALYTICS_URL[chainId as ChainId], label: 'Liquidity', external: true },
-              {
-                link: AGGREGATOR_ANALYTICS_URL,
-                label: 'Aggregator',
-                external: true,
-              },
-            ]}
-          />
-        )}
-        <ExternalNavMenuItem href="https://docs.kyberswap.com">
-          <BookOpen size={14} />
-          <Trans>Docs</Trans>
-        </ExternalNavMenuItem>
-
-        <ExternalNavMenuItem href="https://request.kyberswap.com" onClick={toggle}>
-          <StyledLightIcon />
-          <Trans>Feature Request</Trans>
-        </ExternalNavMenuItem>
-
-        <ExternalNavMenuItem href="https://request.kyberswap.com/roadmap" onClick={toggle}>
-          <StyledRoadMapIcon />
-          <Trans>Roadmap</Trans>
-        </ExternalNavMenuItem>
-
-        <ExternalNavMenuItem href="https://gov.kyber.org">
-          <MessageCircle size={14} />
-          <Trans>Forum</Trans>
-        </ExternalNavMenuItem>
-
-        {under1440 && (
-          <ExternalNavMenuItem href="https://blog.kyberswap.com">
-            <StyledBlogIcon />
-            <Trans>Blog</Trans>
-          </ExternalNavMenuItem>
-        )}
-
-        <ExternalNavMenuItem href="/15022022KyberSwapTermsofUse.pdf">
-          <FileText size={14} />
-          <Trans>Terms</Trans>
-        </ExternalNavMenuItem>
-        {process.env.REACT_APP_MAINNET_ENV !== 'production' && (
-          <NavMenuItem to="/swap-legacy" onClick={toggle}>
-            <Triangle size={14} />
-            <Trans>Swap Legacy</Trans>
-          </NavMenuItem>
-        )}
-
-        <ExternalNavMenuItem href="https://forms.gle/gLiNsi7iUzHws2BY8">
-          <Edit size={14} />
-          <Trans>Contact Us</Trans>
-        </ExternalNavMenuItem>
-
-        <Preferences />
-
-        <ClaimRewardButton
-          disabled={!account || (!!chainId && NETWORKS_INFO[chainId].classic.claimReward === '') || pendingTx}
-          onClick={() => {
-            mixpanelHandler(MIXPANEL_TYPE.CLAIM_REWARDS_INITIATED)
-            toggleClaimPopup()
+      <MenuFlyout node={node} browserCustomStyle={MenuFlyoutBrowserStyle} isOpen={open} toggle={toggle} hasArrow>
+        <Flex
+          sx={{
+            flexDirection: 'column',
+            gap: '16px',
           }}
         >
-          {pendingTx ? (
-            <>
-              <Loader style={{ marginRight: '5px' }} stroke={theme.disableText} /> <Trans>Claiming...</Trans>
-            </>
-          ) : (
-            <Trans>Claim Rewards</Trans>
-          )}
-        </ClaimRewardButton>
-        {!!process.env.REACT_APP_TAG && (
-          <Text
-            fontSize="10px"
-            fontWeight={300}
-            color={theme.subText}
-            mt="16px"
-            textAlign={isMobile ? 'left' : 'center'}
+          <MenuItems />
+
+          <Divider />
+
+          <Preferences />
+
+          <Divider />
+
+          <ClaimRewardButton
+            disabled={!account || (!!chainId && NETWORKS_INFO[chainId].classic.claimReward === '') || pendingTx}
+            onClick={() => {
+              mixpanelHandler(MIXPANEL_TYPE.CLAIM_REWARDS_INITIATED)
+              toggleClaimPopup()
+            }}
           >
-            kyberswap@{process.env.REACT_APP_TAG}
-          </Text>
-        )}
+            {pendingTx ? (
+              <>
+                <Loader style={{ marginRight: '5px' }} stroke={theme.disableText} /> <Trans>Claiming...</Trans>
+              </>
+            ) : (
+              <Trans>Claim Rewards</Trans>
+            )}
+          </ClaimRewardButton>
+          {!!process.env.REACT_APP_TAG && (
+            <Text
+              fontSize="10px"
+              fontWeight={300}
+              color={theme.subText}
+              mt="16px"
+              textAlign={isMobile ? 'left' : 'center'}
+            >
+              kyberswap@{process.env.REACT_APP_TAG}
+            </Text>
+          )}
+        </Flex>
       </MenuFlyout>
       <ClaimRewardModal />
       {chainId && [ChainId.BTTC, ChainId.RINKEBY].includes(chainId) && <FaucetModal />}
     </StyledMenu>
   )
 }
+
+export default Menu
